Make DB path work
This commit takes the db code from previous commits, the api, models,
and schema migrations and fixes the bugs in them to make them work
as intended. From here we have a working base which can be expanded
to make the tool more feature rich.

It should probably be noted that the one test I did at this point
had terrible performance. I loaded a tempest run subunit ouput file
with roughly 2300 tests in it. This might have been either a sqlite
issue or my use of session is far from ideal. I'm assuming the latter.