Add total run time to the runs table
This commit adds calculating the total run time in secs and storing
that in the runs table to the script. This also converts the column
in the runs table to be a float not an int. Normally this would
require a separate migration but since I just pushed this I'm
assuming I'm not going to break existing DBs.