Add metadata tables and cli args for run meta
This commit adds a key value metadata table for each of the 3 object
types: test, run, and test_run including the ORM model for the new
tables. An DB API call is added to set the metadata for a run and
a CLI argument is added to added to set arbitrary metadata about
the run(s) being added to the db.