Some performance improvements to db updates
This commit gets a bit smarter and limits the amount of db writes we
need. Instead of updating everything for every test in the list it
does the summing up front where possible. Also for new tests there is
no need to break up the add and update so this breaks it into an
either or situation. Lastly, the session object is properly being
passed to every operation. Before there were some that were missed.