Update the documentation
This commit updates the documentation for subunit2sql. It adds details
to the README around running setting up the DB and running subunit2sql.
Also, the db api autodoc usage is corrected so that it properly
generates the documentation. Finally, the sphinx theme is switched to
oslosphinx to have consistent style with the documentation for the
other OpenStack projects.

Related-Bug: #1348202
Change-Id: Icf148dee19bb888b551e60a396f87f6c09553e09