Switch sample sqlalchemy url to use mysql
Since the most recent migration broke support for using a sqlite db
it makes sense to update the db in the alembic.ini to use another
db. This commit switches the default url in the alembic.ini to use
a mysql db running on the localhost with the name subunit which is
also the username and password for that database. This is just an
example which should be changed when running the migrations on an
actual db.

Change-Id: Ibf19552758af0818217ab276e0033fcd0b2bce7b