Add new item to the TODO list
Part of subunit2sql's goals is to add support to the testrepository
package to use subunit2sql as a configurable repository type. (just
like the current file, memory, or samba_buildfarm repository types.
Right now the data model of the subunit2sql db schema won't work with
that. Testrespository assumes that runs are added to the repository
in an order and can be retrieved based on that. So in order to support
that use case, support needs to be added to the data model for knowing
the order runs were added to the DB. However since since the schema
will require changes a new migration must be added to add this to the
data model.

This commit updates the TODO list to add this work item.

Change-Id: I1e5f140eb1df21eb1239e02e048346843d3b42ec