Add order column to runs table
This commit adds a timestamp run_at column to the runs table.
This new column is needed to determine the order in which runs are
added to the db. Support for this is needed by testrepository
repositories since it needs to be able to determine which run is the
most recent for some of it's operations. Additionally knowing the
order in which the runs occur is something that is generally useful,
and worth storing in the db.

Change-Id: If3c2251f9c62fc94f2c1716f6091b637c965e610