Add subunit2sql-db-manage utility
This commit adds a new cli utility for managing the subunit2sql db.
Previously alembic could just be used directly however since access
to the db models were needed for future migration scripts the db api
was needed. This required invoking the config object to be able to
use the db api calls. Alembic isn't aware of oslo.db or the config
object, so by creating a seperate cli interface we initialize all of
that at the same time. This also saves the need to configure
connection info to the db in 2 places, since just the oslo.db option
will be used.

This utility borrows heavily from the neutron utility to do the same
thing.

Change-Id: I110baa532d08de4ca70b7ea2d1dcdc845d595693