Preseed data for migration tests until migration 163fd5aa1380
This commits starts preseeding data for tables as they are added for
the initial migrations. Since they are just table adds running checks
wouldn't have any value. The first modification that would be
performed on data in a table occurs in migration 163fd5aa1380, so a
check method is used there.

However, looking at that migration, the new column isn't used, it will
be null for all existing rows so a new migration will need to be added
to do the arithmetic and add data for the existing rows.

Change-Id: I4a874f56f92c65df22923e25908a537eec2cbfdd