Remove unneccesary state_path config option
This commit removes the uncessary state_path config option. This was
only used to manage the location where a sqlite db would be kept.
However, since several migrations don't work with sqlite and support
for using it has been dropped there is no reason to keep the extra
option around. Especially since it's been inconsitently used as the
rest of the project has grown. If sqlite support is ever brought back
this option can be added again in a more uniform and consistent manner.

Change-Id: I88910e85a2248d16bf691a4e01ce5a5f4c64e943