Add migration to populate run_time where needed
The existing migration to add an avg run_time column to the tests table
neglected to calculate the run_time for existing rows and instead just
added a NULL value. Since this migration is part of a released version
of subunit2sql it's too late to add support for adding a run_time to
the columns where it makes sense. This commit adds a new migration
that will add the avg run_time for each test in the tests table where
there are successful test runs and the run_time is still set to NULL.

Additionally there were a couple of fixes around adding migrations
that needed to be added here, mostly just changing the base
alembic.ini to reflect the new relative path of the migration dir and
ignoring an additional pep8 rule which conflicted with the sqlalchemy
syntax.

Change-Id: I0b9c673106c41e9e66a2be5da6ebe9ca58d0ad36