Add cli option to subunit2sql to specify run_id
This commit adds a new cli option to subunit2sql to specify the id to
use for the run being added. This can be used in with a ci system,
like zuul, which already has a unique id to correlate the 2 data sets.
This should be more efficient than storing the external uuid in the
metadata table which was the suggested path prior to this commit.

Change-Id: I5a3e03f1cbf64f4e3ea40ac14bbcb9d1983331f4