Add --average option to sql2subunit cli
This commit adds a new option to the sql2subunit command to generate a
subunit stream using the aggregate data from the tests table. This can
be useful for preseeding a testrepository with data from all previous
runs, which was the intent of the option. When specified instead of a
run_id, the output subunit stream using the test_ids from the tests
tables, the test status will all be success, and timestamps will be
based on the current time and the average run_times. There won't be
any other information in the stream.

Change-Id: I81573b5e38b26fa968ada485cb231194c9231fab