Add new db api methods for getting test data from runs
This commit adds 2 new db api methods get_tests_from_run_id() and
get_tests_run_dicts_from_run_id(). The former will return a list of
models.Test objects for the tests of a given run. The latter is used
to collect the aggregate information across tables for the tests run
during the specified run and return that in a dictionary. This will
be very useful for constructing subunit streams from information in
the database.

Change-Id: I2195e1e7954f3b09ba3b57fab6c276b5aed52b43