Add a db api method to extract run_times for a single test
This commit adds a method to return a time series of test run_times
for a single test which is passed in by the tests id field. The output
dictionary is useful in conjunction with the pandas Series() class,
since it will be recognized as a TimeSeries correctly, which makes
graphing and analysis easier.

Change-Id: I27ed499205d47146b6acfe530ec7ee12865fe965