Share a single connection with all systems during migrations
The purpose of the Alembic env.py file is so that an application
may define the optimal approach to database connectivity
for an application as it runs migrations.  Here, we ensure
that the connectivity comes from the same system as that of
the api.py module itself, e.g. enginefacade, and we also
ensure that the same transaction/connection used for migrations
is shared with the ORM session, so that ORM operations inline
with schema operations share the same transactional context.

Change-Id: Ie2d28d2b04e452c13d16e561b781f82cf2b75e18
Story: #2000099