Fix terrible path finding code in 1679b5bc102 migration
This commit fixes a blatant bug in the 1679b5bc102. If there was a '.'
in the path of the migration file (which there will be if it's
installed, because of python 2.x or python 3.x, then the hack of path
finding code will not generate the correct path and the migration will
fail before it even began.

Change-Id: I4008b13dc356450eb463e81cff7d30230aed96eb