Add modular subunit2sql-graph entry point
This commit replaces the previous entry point in run_time.py for the
subunit2sql-graph command and creates a new module for it in graph.py.
This will enable the expansion of the types of graph commands
supported in future patches, by using an oslo.config subparsers it
will make expanding the subunit2sql-graph command a bit simpler.

Change-Id: I0de9aedae5447690b5433842799758d406b9b6ec