Remove matplotlib from requirements file
This commit removes matplotlib from the requirements file. It is
only used by the subunit2sql-graph tool and isn't required for the
whole package. Given issues with running it in CI this removes it
from the requirements file and adds a note to the README about
needing it to use the graph command.

Change-Id: Ibdc68745ea035be66a35281845d7e4fc49e239a4