Cleanup std dev backfill in run_time graph
This commit makes some improvements to the std dev backfill graphing
on the run_time graph. Previously the raw std dev was just computed
without any respect to what was being displayed and a negative value
would be displayed. However, since it is an elapsed time having a
negative value doesn't make much sense. This filters the generated
std dev time series to floor at 0 to correct this.

Change-Id: I97e89a775112786ff984d5fb00c4ccd8f539448a