Scale run_time graph in subunit2sql-graph docs
This commit adds a missing scale parameter to the run_time graph
image in the subunit2sql-graph docs page. The image is quite large
and not scaling it down makes it comically large and completely
illegible in the docs. While this is very amusing, it is not
exactly useful, so this commit fixes it.

Change-Id: Ic6dfcd133ab9ba59049e3c1b7ae3a30211b7c4e0