Don't try to create indexes twice
In certain environments the index attempted to be created in migration
13d819bbb0ff can sometimes already exist in the db which causes an
error when they already exist. This is likely do to an error on my
part when I originally wrote the migration and not testing things
properly. This commit remedies this situation by checking if the
indexes exist prior to attempting to create them. If an index exists
we just skip the creation step and move on. This is safe since the
end state of the migration doesn't change, it just makes it safe to
execute in more strict DBs.

Change-Id: I901175c2b07efc3e1747e58ca7ea4a46c5149ffa