Set temporary version cap on oslo.db
This commit temporarily adds a cap on oslo.db because of the use of a
private method from oslo.db in the alembic env.py for subunit2sql
which was removed in the latest release. In the future we will switch
to using a public api from oslo.db for doing this. However, one does
not exist at this time. So we add a cap on the version of oslo.db to
ensure that a version including the private api is used.

Change-Id: I726a4afb9b4ed0d7283b42e8f5f45035fcf4fabf