Cleanup dailycount graph subunit2sql-graph command
The code that landed for the dailycount graph command was accidently
an early prototype version of the graph, instead of the more polished
version which was developed. This commit fixes a number of issues with
the command, first the sampling frequency was every second instead of
daily like the graph output implies. This is addressed by resampling
the time series daily. The second issues was that the default fallback
title was "Number of Tempest tests run in the Gate" which is obviously
application specific. Lastly, the date filtering both didn't work and
was using a different format than the help string perscribed. This is
fixed by using the pandas based date filtering in the analysis utils
module instead of the database api call option which uses a different
format.

Change-Id: If0fc077f2f77f6aeea914554d36cb5be8c90f323
Story: #2000311