Add new graph type to show run_time variance by metadata
This commit adds a new graph to subunit2sql-graph. run_time_meta, which
is used to display a scatter matrix of all the aggregate run times for
all runs grouped by a metadata key value. Note depending on the DB size
the plot generation can be fairly computationally and memory intensive.

Change-Id: I369717c84df0217a9a0cf17b7f9e975e580ce5a7