Add a DB API method to get failing test_ids by run metadata
This commit adds a new db api method,
get_failing_ids_from_runs_by_key_value, to get a list of distinct
test_ids that failed during all the runs that match the provided key
value pairs. This is going to potentially used by elastic-recheck in
the future to filter the results of queries.

Change-Id: I13dec0d66740bac966ce447f9cd571016da0f973