Order get_tests_run_dicts_from_run_id response by date
Previously the output from sql2subunit had an arbitrary order of tests
returned. (when not run with --average) While technically not an issue
for the protocol, some programs which analyze the subunit stream (like
subunit-trace) assume tests were added to the stream in chronological
order. To address this issue this commit adds an order_by() call to
the DB query that gets the values to generate the dicts in the
response list. Then to ensure the insertion order is remembered when
the dict is iterated over in sql2subunit later it is switched to an
OrderedDict, which since it's a subclass of dict should retain
backwards compat for any existing code out there. The backwards compat
concern is why the response isn't changed to a list, which would be
the normal way to ensure a series of dicts were treated in order.

Change-Id: Idc21f968f210908f061e663b1d40ddaf7b37204c