Add mysql index length cap on index creation
This commit fixes an issue when running migration 13d819bbb0ff on
mysql when using a  multibyte utf8 character set. The length of the
test_id column on the tests table exceeds the max index lenght in
that environment. This commit sets a max length on the index to avoid
any issues when running it.

Change-Id: I37e1a860a64c09e2da9470104dfda9bb3945f790