Fix using no dates in create_test_run db api method
This commit fixes a minor edge case in using the create_test_run
db api method. If no date was specified, which is an allowed method
to add a test_run, albeit a strange one, the method would stack
trace because the datetime operations performed to normalize the dates
before storage were unconditionally run. This simply adds a check
before hand to ensure things actually exist before we try to manipulate
them.

Change-Id: I1caed74d0aecd164e47adc2d1d552383f7c01612