Fix the filter runs by date helper function
This commit fixes the recently added helper method to filter a runs
query by date to behave correctly. Due to sloppy rushed reviews the
fact that the directions were backwards and the timestamp format were
incorrect landed. This meant spaces were required in the string and
it behved inversly to expectations. This commit fixes these 2 issues.

Change-Id: I42d3372dafaafff9806fca7fe7d56de429dc766d