Fix verbose option inialization
The verbose config option is defined in the set of cli opts defined
for each of the subunit2sql cli commands. When those are called the
options are registered with the global CONF object. However, if the
db api is invoked directly there is no guarantee that the config
option registration has occured. This commit addresses this issue
by changing how the option is created and registered. It is moved
into the db api module to ensure that it always exists before it
is used. The option is then imported into all the cli commands that
were previously registering it.

Change-Id: Id5f45fa8a6d8f2f2f2eebe7e2ac4623cac0f7f10