Add run_time_meta arge to specify values to use in graph
This commit adds a new flag to the run_time_meta command in
subunit2sql-graph. You give a comma seperated list to the cli and
it will only graph the box and whiskers for runs which have metadata
value that is in the list provided. This is useful if you only want
to view a subset of values grouped under a key.

Change-Id: Ic6cadf727ad0421bac93905bc078b2c6a3412b1a