Fix the delete_old_*() db api functions
This commit fixes issues with the 2 delete_old_* functions from the
db api. Previously they were using a join to get a list of metadata
rows to delete as part of the expiration, however this isn't allowed
because sqlalchemy doesn't know which rows to actually delete. This
switches the internals of the function to use a subquery to get a
list of all ids and use that as the filter on the delete command.
Tests are then added to verify the functionality.

Change-Id: Ibe9384cd2b470a10bbc6030dad2bdbb759fee276