Add migration to cleanup and add additional indexes
This commit adds a migration which adds an additional composite index
on id, test_id in the tests table. This is needed to help with the
performance of queries that are being generated by the openstack-health
dashboard's rest api. As part of this migration though some additional
cruft in the test_run indexes is cleaned up. We accidently were ending
up with duplicate indexes on the id columns in the test_runs table.
This removes the duplicate and fixes the check to ensure new DBs
don't end up with duplicate indexes.

Change-Id: I9d4d42aaa6232db0b2ddfaf8d1a5f079cbeb6b31