Add an extension capability for adding targets
This is meant to facilitate anything else that wants to look at the
subunit stream and do something with it.

The unit test for main is simply intended to ensure that found targets
are passed in to the reader. The unit test for ReadSubunit verifies that
any targets passed in are in fact copied into the final result
processor.

Change-Id: I5200521a3c33d61ea4e4af8cfa900eff2c698355
Implements: counter-inspection