Add back uuid column for runs table
When we converted the runs table to use an auto-incrementing integer as
the primary key column we dropped the uuid. However, in doing this we
would be deleting the only user facing identifier for a run. This
commit corrects the oversight by renaming the old id column as a
non-primary key column that can be used for an user input or output
regarding runs. This also adds a new api method and calls it in
sql2subunit to maintain backwards compatibility after the migration.

Change-Id: Idab32bb3449559285f9d45b6132b1c07c1260e7c