Add support to subunit2sql cli to specify a run_at time
This commit adds support to the subunit2sql cli command to specify the
run_at field for a run being created. The db api has had support for
this for quite some time but it was never exposed through the CLI.
This resulted in the default behavior of using utcnow() to set the
run_at column for a created run. However, if you're trying to insert
old runs into the DB this isn't ideal because it will look like the
test results were recent. This commit enables users to specify a
different run_at time for these use cases.

Change-Id: I0c00ef71633e043c9d80adf0bc2df0157d1a11b3