Add migration for adding missing indexes
This commit adds indexes which were previously implicitly added by the
use of foreign keys on the id columns in the metadata tables. However,
after the last migration to convert uuid primary keys to integers the
foreign keys were all dropped. This commit adds these indexes back to
fix the performance issues with queries involving the metadata tables.
Additionally, the index on the run_at column in the runs table was
overlooked, so this migration adds that back as well.

Change-Id: I162b6daec1116937cf499bd56fde92c44590f1d4