Add tests to check migrations and models match
This commit adds a new set of tests to assert that our sqlalchemy ORM
models match the state of the db post migration. Recently we've seen
how far the 2 have drifted apart over time and even after several
patches to correct it they still don't match. This commit corrects
the existing issues as well as adds tests to assert this is correct
moving forward. The model sync tests with sqlite are currently skipped
because the migration sets the id columns to an integer type on sqlite
but the models expect biginteger. This is due to biginteger not
existing on sqlite, so functionally there is no difference but the
tests would fail because of the type mismatch.

Change-Id: I1c25f331fe1143328fef9a7d46fdcec350a037c8