Fix hard coded filter in get_test_run_series()
This commit fixes a hard coded filter made in the db api method
get_test_run_series() which was assuming that the db had run_metadata
like the openstack-infra db. This commit fixes that by adding 2 params
for the key and value that were used to filter. For backwards compat
the defaults were set to the previous hard coded value. However, in a
future release we should remove these defaults since they aren't
correct for every environment.

Change-Id: Ie7daae9dd2d1e8c7dfd5d8913d2e43261597d018