Add release notes for pending 1.2.0 release
This commits all release notes missing for the pending 1.2.0 release.
These items should have been included with the commit that added the
change but was overlooked. We should be better about this in the
future. But, since the ship has sailed already this just ensures we
generate release notes before the release is pushed.

Change-Id: Ia7451e282bc1dbfa627bed0d21b9d8c6c689c5fc