Disable session synchronization on deletes
This commit switches the synchronization_session option on the delete()
calls in the delete_old*() methods in the db api. When setting session
synchronization sqlalchemy is trying to keep a local cache of the data
being deleted. For large DBs like, the infra one, this can eat too
much RAM, which depending on the system might invoke the OOM killer.

Change-Id: Iff5112072fd64d6b96666654edd7cb17897d0765