Add optional run_metadata filter to get_test_runs_by_test_test_id
This commit adds 2 optional kwargs, key and value, to the db api
method get_test_runs_by_test_test_id() to add an optional run_metadata
filter to the method. If both a key and value are specified the
returned test_runs will be filtered on test runs that contain the
specified run_metadata.

Change-Id: I613d5838def780a14b2ecfb4ca54daa72a1c06b2