Remove group regex on .testr.conf
This commit removes the group regex line from the .testr.conf. This
was originally added to aid in test scheduling and decrease lock
contention by ensuring testr always grouped tests in the same class
together. However, since support for multiple backends was added to
all the functional DB tests this has become counterproductive and
actually increases wait time. It in effect was serializing all the
tests which touched a DB regardless of which backend the test uses
because the tests for multiple backends share a class. To avoid this
uneccesary waiting this commit removes the group regex.

Change-Id: Ib38beb2e4fb7b2498c26d495faa349328a7689db