Rework migration 1679b5bc102c to not use the db api
This commit reworks the data migration to split microseconds from
the time column into the separate microseconds column as to not
rely on the DB api. For sqlite and postgres this causes potential
issues when we try to modify the schema for the test_runs table
when this migration is run because the models don't match the state
of the DB when this migration is run. MySQL is unaffected by this
potential issue because it runs the migration separately in
hand written sql.

Change-Id: I527e2ab9d030be639954fdb63584c3c4c6f4dd1c