Enable returning run id in get_test_runs_by_status_for_run_ids()
This commit adds a new option, include_run_id, to the
get_test_runs_by_status_for_run_ids() DB API method for including the
run uuid in the dictionaries returned by the method. This was an
oversight in the method and needed by users in certain cases, however
we can't just add it by default for backwards compat reasons. So a
bool flag is used to enable the additional field being returned.

Change-Id: Icf203668690419c38f6d6be6b6fe4af8462845f3