Fix date format for agg_count and get_test_counts_in_date_range()
This commit fixes an issue with the date formatting on the agg_count
subunit2sql-graph command to be consistent with the other commands
and actually use the ISO8601 format like the docs say it does. However,
because the actual str parsing and format were in the DB API (and for
several releases) we need to gracefully handle the switchover to
anything else. So this commit updates get_test_counts_in_date_range()
to also take in a datetime object and updates the docs to indicate
the previous str parameters are deprecated. It then updates the
agg_count command to own it's own string parsing and just pass a
datetime object to the DB api.

Change-Id: Ie7c67941e59c70285263611803f73b23b69ee06b
Story: #2000581