Move oslo.concurrency to test-requirements
Looking through the subunit2sql code oslo.concurrency is only ever used
for a lock fixture used during the unit tests. We shouldn't have test
only requirements listed in the requirements.txt file. This commit moves
it to live in test-requirements so we don't unecessarily install it
when installing subunit2sql.

Change-Id: Idf537d20980ffc1c1d3ec8c7758c810f18fc5164