Add option to ensure get_test_runs_by_test_id() is ordered by date
This commit adds a new option to the get_test_runs_by_test_id() DB API
function, most_recent_first, that when set True will order the response
list by date with the most recent being first. This is useful if you
want to look at a recent subset of the response list on the consumer
side.

Change-Id: Ia94f8066eb36deb33391162920b748707630e420