Remove the test_attr_prefix before injecting
The test_attr_prefix allows filtering test attributes for injection
into the test_metadata table. Such attributes only get a prefix for
the sake of filtering, so the prefix should be removed before they
are injected into the DB.

This is a change in behaviour, but I believe it's safe to do so,
since the CLI option was broken until now, which suggests no-one was
using it really.

Change-Id: I0cf3701a143f6fb71d03a0a7bdfb0a84c107ecbe