Add DB API function to get multiple tests by test_id
This commit adds a new DB API function, get_tests_by_test_ids(), to get
multiple tests by test_id at once. There is already a function to do
this for an individual test, but if you need to get multiple tests by
tests looping that results in a lot of extra DB traffic when it can be
done in a single request.

Change-Id: I9db6c050720bd12693ff4fa29ee4f43fe60169be