Add run_failure_rate argument to specify values to use in graph
This This commit adds a new flag to the run_failure_rate command in
subunit2sql-graph. It expects a comma separated list on the cli and it
will only graph the bar for runs which have a metadata value that is in
the list provided. This is useful if you only want to view a subset of
values grouped under a key.

Change-Id: I70b7d34bd027ee2a6f7e75fc0cef5a39ddcda3ef