Fix dailycount command
This commit fixes the dailycount subunit2sql-graph command. This was one
of the earlier graph commands added and has since bitrotted a bit to the
point where it doesn't work anymore. This commit fixes all those issues
and bumps the minimum versions for pandas and matplotlib because of the
updated usage.

Story: 2001034
Task: 4590
Change-Id: Iaea5e2f6036d39b6aa57685d484cd02f323a4d39