Fix unit using incorrect type in fake default
This commit fixes a bug in the subunit2sql shell unit tests. The test
sets a fake default value to check it gets set properly later in the
tests. However, newer versions of oslo config do type checking and this
option was expecting a dict, not a string. This commit fixes this by
setting it the fake value to be a dict so we don't fail in oslo.config
anymore.

Change-Id: I5830cb271e5c0c20637323b52fced1483209ec96