Add subunit2sql CLI option to use non_subunit_name
This commit adds a new CLI option for allowing mixed subunit
content. This is desired for subunit.stream files with non-subunit
content mixed in but still want the subunit content converted to
sql. However, this commit only adds support for passing the argument
to ByteStreamToStreamResult. Nothing is added beyond that for managing
the non-subunit content.

Co-Authored-By: Matthew Treinish <mtreinish@kortar.org>

Change-Id: I68cbeb47e9093b98630f096fc5818a77932c8da4