Switch to use stestr completely
This commit adds .stestr to .gitignore and remove os-testr from
test-requirements.txt to switch to use stestr completely. We actually
already switched in the previous commit[1]. And now, we can switch
completely.

One thing - .testrepository is still remaining in .gitignore. This
directory can be still remaining in the most of user environments. So
that, to avoid the confusion, '.testrepository' should be still there
for a moment.

[1] I39f5db8f77de027c18917215c761dca69d8cbea0

Change-Id: I9530af1a91ae20dca24a699fbf04893102570f09