Fix mysql db api
By default the multiline option for the db api was included in
PyMySql<0.8 but now that it is removed[1] we need to pass in the
multi=True argument to the cursor to maintain the same execution
behavior. Unfortunately, I wasn't sure on how to do that
so I modified the op.execute method to run the sql commands one
line at a time instead. It seems to work for the test cases.

[1]https://github.com/PyMySQL/PyMySQL/commit/c0aa3179406571592d3beb9d5a35badc4047cb79

Change-Id: If139f0941562b7a792d813b5e23a47eb92487304
Story: 2001420