Add --dpi flag to subunit2sql-graph
This commit adds a new flag, --dpi, to the subunit2sql-graph command.
This flag is used to specify a DPI for generated graph images.
Previously most graph commands hard coded the DPI to 900 (which is far
too high for most people) and others didn't specify one. This commit
changes that so it's user adjustable and consistent between all
commands.

Change-Id: I5c8c48bda18daff11d6069c35451d932ffe4e33b