Honor dpi cli flag for ci_time graph
The newly added ci_time graph was originally introduced before the dpi
cli flag for the subunit2sql-graph command. It merged around the same
time as the cli flag meaning it was never updated to respect that, and
had a hardcoded dpi of 900. This commit fixes that oversight and updates
the command to use the dpi flag.

Change-Id: I72b3dc5d4d484a2b9f7b542ab1da01cc40a0ed24