Fix pandas rolling calls in run_time graph
This commit fixes the use of rolling_mean() and rolling_std() in the
run_time graph command. These functions were removed from pandas a few
releases ago and were replaced by the Rolling class. We updated the
other commands already, but the run_time graph was missed. This commit
corrects the oversight and uses the Rolling class to generate the
rolling mean and std dev data.

Change-Id: I4a2fd0256b05f468cdbbe67fb2573233f64d74d8