Clean up run_time graph style
This commit cleans up the style used for the run_time graph. It switches
the individual run_time graph (not the avg or std dev) from a line graph
to scatter points. Since there are potentially a lot of data points this
makes things cleaner and easier to see. Then the size of those points
and the line width are shrunk. When there are a lot of data points
things overlap, get cluttered, and hard to see. By shrinking the lines
and scatter points it alleviates this to a certain degree.

Change-Id: I1271acf10a12df8e36a85e4deacbfc77f7f05bc0