Fix handling of xfail and uxsuccess results in subunit2sql cli
This commit fixes an issue with the subunit2sql cli (for importing
subunit streams to the DB) when the subunit stream has a test_run with
an xfail or unxsuccess status. Previously we didn't handle these
statuses at all and we'd raise an exception because they weren't known.
This fixes this by treating xfail the same as success and uxsuccess the
same as failure.

Change-Id: Ib84eb5e4b2da4847e86ff24c8bc5ba19a0df5b37