Fix inconsistent treatment of test status
This commit is the follow on to change
Ib84eb5e4b2da4847e86ff24c8bc5ba19a0df5b37 which fixed an unhandled
condition when xfail or uxsuccess was in the passed in results stream.
While that fixed the failure condition it neglected to update how we
treat those test result statuses in all other cases. This commit does
that by making sure we treat xfail and uxsuccess as success and failure
respectively for aggregate counts.

Change-Id: I1d2308d2c68f4f6b432ab733ead19a0167ba0b94
Story: 2000701
Task: 3176